Add staff package endpoint
